pyproject.toml
[project]
name = "clarity"
version = "0.1.0"
description = ""
authors = [
    {name = "Invictus808",email = "contact@invictus808.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client (>=2.160.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
